+++ weblate/checks/placeholders.py
@@ -22,6 +22,7 @@ import re
 
 from django.utils.html import escape
 from django.utils.safestring import mark_safe
+from django.utils.translation import gettext
 from django.utils.translation import gettext_lazy as _
 
 from weblate.checks.base import TargetCheckParametrized
@@ -29,6 +30,8 @@ from weblate.checks.parser import multi_value_flag, single_value_flag
 
 
 def parse_regex(val):
+    if isinstance(val, re.Pattern):
+        return val
     return re.compile(val)
 
 
@@ -36,39 +39,67 @@ class PlaceholderCheck(TargetCheckParametrized):
     check_id = "placeholders"
     default_disabled = True
     name = _("Placeholders")
-    description = _("Translation is missing some placeholders:")
+    description = _("Translation is missing some placeholders")
 
     @property
     def param_type(self):
-        return multi_value_flag(str)
+        return multi_value_flag(lambda x: x)
+
+    def get_value(self, unit):
+        return re.compile(
+            "|".join(
+                param.pattern if isinstance(param, re.Pattern) else re.escape(param)
+                for param in super().get_value(unit)
+            )
+        )
 
     def check_target_params(self, sources, targets, unit, value):
-        return any(any(param not in target for param in value) for target in targets)
+        expected = set(value.findall(unit.source_string))
+
+        missing = set()
+        extra = set()
+
+        for target in targets:
+            found = set(value.findall(target))
+            missing.update(expected - found)
+            extra.update(found - expected)
+
+        if missing or extra:
+            return {"missing": missing, "extra": extra}
+        return False
 
     def check_highlight(self, source, unit):
         if self.should_skip(unit):
             return []
         ret = []
 
-        regexp = "|".join(re.escape(param) for param in self.get_value(unit))
+        regexp = self.get_value(unit)
 
-        for match in re.finditer(regexp, source):
+        for match in regexp.finditer(source):
             ret.append((match.start(), match.end(), match.group()))
         return ret
 
     def get_description(self, check_obj):
         unit = check_obj.unit
-        if not self.has_value(unit):
-            return super().get_description(check_obj)
-        targets = unit.get_target_plurals()
-        missing = [
-            param
-            for param in self.get_value(unit)
-            if any(param not in target for target in targets)
-        ]
-        return mark_safe(
-            "{} {}".format(escape(self.description), escape(", ".join(missing)))
+        result = self.check_target_unit(
+            unit.get_source_plurals(), unit.get_target_plurals(), unit
         )
+        if not result:
+            return super().get_description(check_obj)
+
+        errors = []
+        if result["missing"]:
+            errors.append(
+                gettext("Following format strings are missing: %s")
+                % ", ".join(sorted(result["missing"]))
+            )
+        if result["extra"]:
+            errors.append(
+                gettext("Following format strings are extra: %s")
+                % ", ".join(sorted(result["extra"]))
+            )
+
+        return mark_safe("<br />".join(escape(error) for error in errors))
 
 
 class RegexCheck(TargetCheckParametrized):